# openapi/web-v1.json
{
  "components": {
    "schemas": {
      "ActivityResponse": {
        "properties": {
          "by_day": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Day",
            "type": "object"
          },
          "by_project": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Project",
            "type": "object"
          },
          "by_tool": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Tool",
            "type": "object"
          },
          "degradation_rate": {
            "title": "Degradation Rate",
            "type": "number"
          },
          "degraded_calls": {
            "title": "Degraded Calls",
            "type": "integer"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "failed_calls": {
            "title": "Failed Calls",
            "type": "integer"
          },
          "failure_rate": {
            "title": "Failure Rate",
            "type": "number"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "total_calls": {
            "title": "Total Calls",
            "type": "integer"
          },
          "total_results": {
            "title": "Total Results",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end",
          "total_calls",
          "total_results",
          "failed_calls",
          "failure_rate",
          "degraded_calls",
          "degradation_rate",
          "by_day",
          "by_tool",
          "by_project"
        ],
        "title": "ActivityResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "recallum_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Recallum Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me/activity": {
      "get": {
        "operationId": "own_activity_me_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Own Activity",
        "tags": [
          "self-service"
        ]
      }
    },
    "/me/api-keys": {
      "get": {
        "operationId": "list_api_keys_me_api_keys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  },
                  "title": "Response List Api Keys Me Api Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Api Keys",
        "tags": [
          "self-service"
        ]
      },
      "post": {
        "operationId": "issue_api_key_me_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedApiKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Issue Api Key",
        "tags": [
          "self-service"
        ]
      }
    },
    "/me/api-keys/{key_id}": {
      "delete": {
        "operationId": "revoke_api_key_me_api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Key Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Revoke Api Key",
        "tags": [
          "self-service"
        ]
      }
    },
    "/me/memories": {
      "get": {
        "operationId": "list_memories_me_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "global",
                    "project"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "preference",
                    "decision",
                    "constraint",
                    "fact"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "stale",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stale"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Memories",
        "tags": [
          "self-service"
        ]
      },
      "post": {
        "operationId": "create_memory_me_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RememberResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Memory",
        "tags": [
          "self-service"
        ]
      }
    },
    "/me/memories/reassign-project": {
      "post": {
        "operationId": "reassign_project_me_memories_reassign_project_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReassignResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Reassign Project",
        "tags": [
          "self-service"
        ]
      }
    },
...
    "/me/memories/{memory_id}/history": {
      "get": {
        "operationId": "memory_history_me_memories__memory_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Memory History",
        "tags": [
          "self-service"
        ]
      }
    },
    "/me/memories/{memory_id}/supersede": {
      "post": {
        "operationId": "supersede_memory_me_memories__memory_id__supersede_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupersedeMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupersedeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Supersede Memory",
        "tags": [
          "self-service"
        ]
      }
    },
...
    "/me/memory-profile": {
      "get": {
        "operationId": "memory_profile_me_memory_profile_get",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileBlock"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Memory Profile",
        "tags": [
          "self-service"
        ]
      }
    },
...
  }
}